Move test files under controller folder
