+++ framework/test/jest/specs/unit/controller/schema/config.spec.js
@@ -1,4 +1,4 @@
-const config = require('../../../../../src/controller/schema/config');
+const config = require('../../../../../../src/controller/schema/config');
 
 describe('schema/config.js', () => {
 	it('schema config must match to the snapshot.', () => {
